requirements.txt
numpy==2.1.3
pandas==2.2.3
jupyter==1.0.0
requests==2.31.0
kafka-python==2.0.2
confluent-kafka==2.3.0
sseclient-py==1.8.0
psycopg2-binary==2.9.10
